requirements.txt
flask==2.3.3
Pillow==10.0.1
python-dotenv==1.0.0
requests==2.31.0
pydub==0.25.1
numpy==1.24.3
openai==1.12.0
